Adding dependency on new separate library repository. Removed code from this project, moved it to the library project
